+++ src/steps/storage/index.test.ts
@@ -55,6 +55,7 @@ describe('#fetchCloudStorageBuckets', () => {
           retentionPeriod: { type: 'string' },
           retentionDate: { type: 'string' },
           public: { type: 'boolean' },
+          isSubjectToObjectAcls: { type: 'boolean' },
           classification: { const: null },
           etag: { type: 'string' },
         },